+++ changelog.md
@@ -22,6 +22,7 @@ Codion Change Log
 - TableConditionModel.persist() added, for excluding conditions when clearing the model.
 - MultiItemSelection.Index added.
 - FilterModel.Refresher.observable() renamed active().
+- ConditionModel.Operands.lowerBound() and upperBound() shortened to lower() and upper().
 ### is.codion.common.rmi
 - ServerConfiguration.serverName() now throws exception if the server name supplier returns null or an empty string.
 - Server.serverLoad() removed, related refactoring.